scraping/pyproject.toml
[project]
name = "watclub-scraping"
version = "0.1.0"
description = "Web scraping for University of Waterloo student organizations"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "playwright>=1.40.0",
    "beautifulsoup4>=4.12.0",
    "requests>=2.31.0",
    "pandas>=2.1.0",
    "pydantic>=2.5.0",
    "aiohttp>=3.9.0",
    "python-dotenv>=1.0.0",
    "loguru>=0.7.0",
    "httpx>=0.25.0",
    "tenacity>=8.2.0", # For retry logic
    "openai>=1.100.2",
    "chromium>=0.0.0",
    "aiofiles>=24.1.0",
    "instaloader>=4.14.2",
    "duckduckgo-search>=8.1.1",
    "ddgs>=9.5.4",
    "google-api-python-client>=2.179.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
]

[tool.black]
line-length = 88
target-version = ['py311']

[tool.ruff]
line-length = 88
target-version = "py311"
select = ["E", "F", "W", "C90", "I", "N", "UP", "YTT", "S", "BLE", "FBT", "B", "A", "COM", "C4", "DTZ", "T10", "EM", "EXE", "FA", "ISC", "ICN", "G", "INP", "PIE", "T20", "PYI", "PT", "Q", "RSE", "RET", "SLF", "SLOT", "SIM", "TID", "TCH", "INT", "ARG", "PTH", "ERA", "PD", "PGH", "PL", "TRY", "FLY", "NPY", "AIR", "PERF", "FURB", "LOG", "RUF"]
ignore = ["E501", "S101", "PLR0913"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
